BID-91 홈화면 서비스 레이어 분리 및 리팩토링
