+++ Bidit/Bidit/DI/RepositoryProvider.swift
@@ -12,11 +12,19 @@ import Foundation
 final class RepositoryProvider {
     // singleton repositories
     static var userRepo:UserRepository? = nil
+    static var itemRepo:ItemRepository? = nil
     
     static func getUserRepository() -> UserRepository {
         if self.userRepo == nil {
             self.userRepo = UserRepositoryImpl(apollo: Network.shared.apollo)
         }
         return self.userRepo!
+    }
+    
+    static func getItemRepository() -> ItemRepository{
+        if self.itemRepo == nil {
+            self.itemRepo = ItemRepositoryImpl(apollo: Network.shared.apollo)
+        }
+        return self.itemRepo!
     }
 }
