+++ Bidit/Bidit/presentation/Views/homeScene/EndingSoonList/Cell/EndingSoonCell.swift
